+++ e2e/wallets/integration-wallet-package-keepkey/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @coinmasters/integration-wallet-keepkey
 
+## 0.5.36
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@13.0.36
+  - @coinmasters/types@4.7.36
+  - @coinmasters/core@11.0.36
+
 ## 0.5.35
 
 ### Patch Changes