+++ packages/pioneer/pioneer-db-sql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @coinmasters/pioneer-sdk
 
+## 4.7.35
+
+### Patch Changes
+
+- bump
+
 ## 4.7.34
 
 ### Patch Changes