+++ packages/toolboxes/cosmos/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @coinmasters/toolbox-cosmos
 
+## 11.0.36
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.36
+  - @coinmasters/api@3.7.36
+
 ## 11.0.35
 
 ### Patch Changes