+++ packages/toolboxes/evm/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coinmasters/toolbox-evm
 
+## 11.0.36
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.36
+
 ## 11.0.35
 
 ### Patch Changes