+++ packages/wallets/evm-extensions/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @coinmasters/wallet-evm-extensions
 
+## 11.0.36
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.7.36
+  - @coinmasters/toolbox-evm@11.0.36
+
 ## 11.0.35
 
 ### Patch Changes